+++ R/loo-helpers.R
@@ -297,6 +297,8 @@ compare_ic <- function(..., x = NULL, ic = c("loo", "waic", "kfold")) {
 #'   \code{"marglik"} (log marginal likelihood).
 #' @param model_name Optional name of the model. If \code{NULL}
 #'   (the default) the name is taken from the call to \code{x}.
+#' @param overwrite Logical; Indicates if already stored fit 
+#'   indices should be overwritten. Defaults to \code{FALSE}.
 #' @param ... Further arguments passed to the underlying 
 #'   functions computing the information criteria or fit indices.
 #'   
@@ -322,7 +324,8 @@ add_ic <- function(x, ...) {
 
 #' @rdname add_ic
 #' @export
-add_ic.brmsfit <- function(x, ic = "loo", model_name = NULL, ...) {
+add_ic.brmsfit <- function(x, ic = "loo", model_name = NULL, 
+                           overwrite = FALSE, ...) {
   unused_args <- intersect(names(list(...)), args_not_for_reloo())
   if (length(unused_args)) {
     unused_args <- collapse_comma(unused_args)
@@ -338,6 +341,10 @@ add_ic.brmsfit <- function(x, ic = "loo", model_name = NULL, ...) {
   if (!length(ic) || !all(ic %in% valid_ics)) {
     stop2("Argument 'ic' should be a subset of ",
           collapse_comma(valid_ics))
+  }
+  overwrite <- as_one_logical(overwrite)
+  if (overwrite) {
+    x[ic] <- list(NULL)
   }
   args <- list(x, ...)
   for (fun in intersect(ic, c("loo", "waic", "kfold"))) {